Rename the warnings spoke to warnings_spoke (#1014220)
It was shadowing the Python standard library warnings module.
The py-compile script used by autotools to generate bytecode for
Python modules is stupid enough to *prepend* its PWD to PYTHONPATH
and because of that it trying to import the warnings spoke in place of the
warnings module, breaking the Anaconda package build.